requirements.txt
aiohttp==3.6.2
aniso8601==7.0.0
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.2.1
Automat==0.8.0
certifi==2019.11.28
cffi==1.14.0
channels==2.4.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
Django==2.2.10
django-channels-graphql-ws==0.4.2
django-cors-headers==3.2.1
django-environ==0.4.5
django-graphql-jwt==0.3.0
graphene==2.1.8
graphene-django==2.8.0
graphql-core==2.2.1
graphql-relay==2.0.1
gunicorn==20.0.4
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
msgpack==0.6.2
multidict==4.7.4
oauthlib==3.1.0
promise==2.3
psycopg2==2.8.4
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyHamcrest==2.0.0
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dotenv==0.11.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
Rx==1.6.1
service-identity==18.1.0
singledispatch==3.4.0.3
six==1.14.0
sqlparse==0.3.0
Twisted==19.10.0
txaio==20.1.1
urllib3==1.25.8
yarl==1.4.2
zope.interface==4.7.1
